Reconceptualise replica drivers as doc drivers
