+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next
 
+- BREAKING: IReplicaDriver is now used for a class with both `docDriver` and
+  `blobDriver` properties
+- BREAKING: IReplicaDriver is now IReplicaDocDriver, classes have changed names
+  so ReplicaDriverMemory is now DocDriverMemory
 - Improvement: noble-ed25519 has been updated, and is now audited and 10 - 15%
   faster.
 - BREAKING: `Peer` no longer has a `peerId` property.
