+++ src/entries/deno.ts
@@ -1,5 +1,5 @@
-export { ReplicaDriverLocalStorage } from "../replica/replica-driver-local-storage.ts";
-export { ReplicaDriverSqlite } from "../replica/replica-driver-sqlite.deno.ts";
+export { DocDriverLocalStorage } from "../replica/doc_drivers/localstorage.ts";
+export { DocDriverSqlite } from "../replica/doc_drivers/sqlite.deno.ts";
 export { syncReplicaAndFsDir } from "../sync-fs/sync-fs.ts";
 export { CryptoDriverSodium } from "../crypto/crypto-driver-sodium.ts";
-export { ReplicaDriverSqliteFfi } from "../replica/replica_driver_sqlite_ffi.ts";
+export { DocDriverSqliteFfi } from "../replica/doc_drivers/sqlite_ffi.ts";
