+++ src/replica/doc_drivers/sqlite.deno.ts
@@ -1,11 +1,11 @@
-import { ShareAddress } from "../util/doc-types.ts";
+import { ShareAddress } from "../../util/doc-types.ts";
 import {
   EarthstarError,
   isErr,
   ReplicaIsClosedError,
   ValidationError,
-} from "../util/errors.ts";
-import { CoreDoc, IReplicaDriver } from "./replica-types.ts";
+} from "../../util/errors.ts";
+import { CoreDoc, IReplicaDocDriver } from "../replica-types.ts";
 import {
   CREATE_CONFIG_TABLE_QUERY,
   CREATE_DOCS_TABLE_QUERY,
@@ -22,22 +22,28 @@ import {
   SET_ENCODING_QUERY,
   UPSERT_CONFIG_QUERY,
   UPSERT_DOC_QUERY,
-} from "./replica-driver-sqlite.shared.ts";
-import * as Sqlite from "https://deno.land/x/sqlite3@0.4.2/mod.ts";
+} from "./sqlite.shared.ts";
+import * as Sqlite from "https://deno.land/x/sqlite@v3.2.0/mod.ts";
 
 //--------------------------------------------------
 
-import { Logger } from "../util/log.ts";
-import { bytesToString, stringToBytes } from "../util/bytes.ts";
-import { Query } from "../query/query-types.ts";
-import { cleanUpQuery } from "../query/query.ts";
-import { sortedInPlace } from "./compare.ts";
-import { checkShareIsValid } from "../core-validators/addresses.ts";
-import { DocEs4 } from "../formatters/formatter_es4.ts";
+import { Logger } from "../../util/log.ts";
+import { bytesToString, stringToBytes } from "../../util/bytes.ts";
+import { Query } from "../../query/query-types.ts";
+import { cleanUpQuery } from "../../query/query.ts";
+import { sortedInPlace } from "../compare.ts";
+import { checkShareIsValid } from "../../core-validators/addresses.ts";
+import { DocEs4, FormatterEs4 } from "../../formatters/formatter_es4.ts";
+import { ExtractDocType } from "../../formatters/formatter_types.ts";
 
 const logger = new Logger("storage driver sqlite node", "yellow");
 
-type SqlDoc = {
+interface ConfigObject extends Sqlite.RowObject {
+  key: string;
+  content: string;
+}
+
+interface DocObject extends Sqlite.RowObject {
   format: string;
   workspace: string;
   path: string;
@@ -49,17 +55,14 @@ type SqlDoc = {
   signature: string;
   localIndex?: number;
   toSortWithinPath?: number;
-};
+}
 
-/** A strorage driver which persists to SQLite using native bindings. Works in Deno.
- *
- * *Requires the `--unstable` flag to be passed to Deno when used.*
- */
-export class ReplicaDriverSqliteFfi implements IReplicaDriver {
+/** A strorage driver which persists to SQLite. Works in Deno and browsers. */
+export class DocDriverSqlite implements IReplicaDocDriver {
   share: ShareAddress;
   _filename: string;
   _isClosed = false;
-  _db: Sqlite.Database = null as unknown as Sqlite.Database;
+  _db: Sqlite.DB = null as unknown as Sqlite.DB;
   _maxLocalIndex: number;
 
   //--------------------------------------------------
@@ -81,18 +84,6 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
       } catch (err) {
         logger.error("Failed to delete Sqlite file.");
         logger.error(err);
-      }
-
-      try {
-        await Deno.remove(`${this._filename}-shm`);
-      } catch {
-        // If it's not there, that's fine.
-      }
-
-      try {
-        await Deno.remove(`${this._filename}-wal`);
-      } catch {
-        // If it's not there, that's fine.
       }
     }
     this._isClosed = true;
@@ -162,21 +153,22 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
       throw addressIsValidResult;
     }
 
-    this._db = new Sqlite.Database(this._filename, {
+    this._db = new Sqlite.DB(this._filename, {
       memory: this._filename === ":memory:",
     });
     this._ensureTables();
 
-    const [maxLocalIndexResult] = this._db.queryObject<{
-      "MAX(localIndex)": number | null;
-    }>(
+    const maxLocalIndexQuery = this._db.prepareQuery<[number]>(
       MAX_LOCAL_INDEX_QUERY,
     );
 
-    const maxLocalIndex = maxLocalIndexResult["MAX(localIndex)"];
+    const [maxLocalIndexFromDb] = maxLocalIndexQuery.one();
+    maxLocalIndexQuery.finalize();
 
     // We have to do this because the maxLocalIndexDb could be 0, which is falsy.
-    this._maxLocalIndex = maxLocalIndex !== null ? maxLocalIndex : -1;
+    this._maxLocalIndex = maxLocalIndexFromDb !== null
+      ? maxLocalIndexFromDb
+      : -1;
 
     // check share
     if (opts.mode === "create") {
@@ -262,26 +254,40 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
     if (this._isClosed) {
       throw new ReplicaIsClosedError();
     }
-    this._db.execute(UPSERT_CONFIG_QUERY, { key: key, content: content });
+    this._db.query(UPSERT_CONFIG_QUERY, { key: key, content: content });
 
     return Promise.resolve();
   }
 
   _getConfigSync(key: string): string | undefined {
-    const [configQueryResult] = this._db.queryObject<{ content: string }>(
+    const configQuery = this._db.prepareQuery<Sqlite.Row, ConfigObject>(
       SELECT_CONFIG_CONTENT_QUERY,
-      { key },
     );
 
-    if (configQueryResult) {
-      return configQueryResult.content;
+    try {
+      const row = configQuery.oneEntry({ key });
+      const result = row.content;
+
+      logger.debug(
+        `getConfig(${JSON.stringify(key)}) = ${JSON.stringify(result)}`,
+      );
+
+      return result;
+    } catch {
+      return undefined;
+    } finally {
+      configQuery.finalize();
     }
   }
 
   _listConfigKeysSync(): string[] {
-    const keysQuery = this._db.queryArray<string[]>(SELECT_KEY_CONFIG_QUERY);
+    const keysQuery = this._db.prepareQuery<string[]>(SELECT_KEY_CONFIG_QUERY);
 
-    return sortedInPlace(keysQuery.map(([key]) => key));
+    const rows = keysQuery.all();
+
+    keysQuery.finalize();
+
+    return sortedInPlace(rows.map(([key]) => key));
   }
 
   getConfig(key: string): Promise<string | undefined> {
@@ -304,7 +310,7 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
       throw new ReplicaIsClosedError();
     }
 
-    this._db.execute(DELETE_CONFIG_QUERY, { key: key });
+    this._db.query(DELETE_CONFIG_QUERY, { key: key });
 
     return Promise.resolve(this._db.changes > 0);
   }
@@ -340,7 +346,9 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
     logger.debug("  sql:", sql);
     logger.debug("  params:", params);
 
-    const docs = this._db.queryObject<SqlDoc>(sql, params);
+    const docsQuery = this._db.prepareQuery<Sqlite.Row, DocObject>(sql);
+
+    const docs = docsQuery.allEntries(params);
 
     if (query.historyMode === "latest") {
       // remove extra field we added to find the winner within each path
@@ -363,13 +371,14 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
     docsWithStringContent.forEach((doc) => Object.freeze(doc));
     logger.debug(`  result: ${docs.length} docs`);
 
+    docsQuery.finalize();
     return Promise.resolve(docsWithStringContent as DocEs4[]);
   }
 
   //--------------------------------------------------
   // SET
 
-  upsert<DocType extends CoreDoc>(
+  upsert<DocType extends ExtractDocType<typeof FormatterEs4>>(
     doc: DocType,
   ): Promise<DocType> {
     // Insert new doc, replacing old doc if there is one
@@ -395,7 +404,7 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
     };
 
     //  TODOM3: Fix this any type.
-    this._db.execute(UPSERT_DOC_QUERY, docWithBytes as any);
+    this._db.query(UPSERT_DOC_QUERY, docWithBytes as any);
 
     return Promise.resolve(docWithLocalIndex);
   }
@@ -407,10 +416,11 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
 
     const now = Date.now() * 1000;
 
-    const docsToDelete = this._db.queryObject<SqlDoc>(
+    const toDeleteQuery = this._db.prepareQuery<Sqlite.Row, DocObject>(
       SELECT_EXPIRED_DOC_QUERY,
-      { now },
     );
+
+    const docsToDelete = toDeleteQuery.allEntries({ now });
 
     const docsWithStringContent = docsToDelete.map((docToDelete) => ({
       ...docToDelete,
@@ -421,7 +431,9 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
     docsWithStringContent.forEach((doc) => delete doc.localIndex);
     docsWithStringContent.forEach((doc) => Object.freeze(doc));
 
-    this._db.execute(DELETE_EXPIRED_DOC_QUERY, { now });
+    toDeleteQuery.finalize();
+
+    this._db.query(DELETE_EXPIRED_DOC_QUERY, { now });
 
     return Promise.resolve(docsWithStringContent as CoreDoc[]);
   }
@@ -439,24 +451,22 @@ export class ReplicaDriverSqliteFfi implements IReplicaDriver {
     }
 
     // make sure sqlite is using utf-8
-    this._db.execute(SET_ENCODING_QUERY);
-    const [{ encoding }] = this._db.queryObject(GET_ENCODING_QUERY);
+    this._db.query(SET_ENCODING_QUERY);
+    const encoding = this._db.query(GET_ENCODING_QUERY);
 
-    if (encoding !== "UTF-8") {
-      throw new Error(
-        `sqlite encoding is stubbornly set to ${encoding} instead of UTF-8`,
-      );
-    }
+    /*
+        if (res !== "UTF-8") {
+            throw new Error(
+                `sqlite encoding is stubbornly set to ${encoding} instead of UTF-8`,
+            );
+        }*/
 
-    this._db.execute(CREATE_DOCS_TABLE_QUERY);
-    this._db.execute("pragma journal_mode = WAL");
-    this._db.execute("pragma synchronous = normal");
-    this._db.execute("pragma temp_store = memory");
-    this._db.execute(CREATE_LOCAL_INDEX_INDEX_QUERY);
+    this._db.query(CREATE_DOCS_TABLE_QUERY);
+    this._db.query(CREATE_LOCAL_INDEX_INDEX_QUERY);
 
     // the config table is used to store these variables:
     //     share - the share this store was created for
     //     schemaVersion
-    this._db.execute(CREATE_CONFIG_TABLE_QUERY);
+    this._db.query(CREATE_CONFIG_TABLE_QUERY);
   }
 }
